requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pip==25.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
SQLAlchemy==2.0.47
starlette==0.52.1
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
psycopg2-binary==2.9.10